@@ -1,11 +1,9 @@
 [project]
 name = "auto-lazy-imports"
-authors = [
-    { name = "Dhia Hmila" },
-]
+authors = [{ name = "Dhia Hmila" }]
 description = "A module to enable lazy imports using native python syntax"
 license = { text = "MIT License" }
-requires-python = ">=3.9"
+requires-python = ">=3.10"
 keywords = []
 readme = "README.md"
 dynamic = ['version']
@@ -44,48 +39,60 @@ code = "__import__('lazyimports').install()"
 [tool.hatch.version]
 path = "src/lazyimports/__init__.py"
 
+[tool.hatch.build.targets.sdist]
+exclude = [
+    "tests/",
+    ".github/",
+    "uv.lock",
+    ".pre-commit-config.yaml",
+    ".copier-answers.yml",
+    ".gitignore",
+]
+
 [tool.hatch.build.targets.wheel]
 packages = ["src/lazyimports"]
-sources = ["src"]
+
+[tool.hatch.build.targets.wheel.sources]
+"src/lazyimports" = "lazyimports"
 
 
 [tool.ruff]
-target-version = "py39"
+target-version = "py310"
 
 [tool.ruff.lint]
 select = [
-    "A", # flake8-builtins
-    "B", # flake8-bugbear
-    "E", # pycodestyle - Error
-    "W", # pycodestyle - Warning
-    "F", # Pyflakes
-    "N", # pep8-naming
-    "S", # flake8-bandit
+    "A",   # flake8-builtins
+    "B",   # flake8-bugbear
+    "E",   # pycodestyle - Error
+    "W",   # pycodestyle - Warning
+    "F",   # Pyflakes
+    "N",   # pep8-naming
+    "S",   # flake8-bandit
     "RUF", # Ruff-specific rules
     "PLC", # Pylint Convention
     "PLE", # Pylint Error
     "PLW", # Pylint Warning
     "PLR", # Pylint Refactor
-    "Q", # flake8-quotes
+    "Q",   # flake8-quotes
 ]
 extend-select = [
-    "ARG", # flake8-unused-arguments
+    "ARG",   # flake8-unused-arguments
     "ASYNC", # flake8-async
-    "BLE", # flake8-blind-except
-    "C4", # flake8-comprehensions
-    "C90", # mccabe (complexity)
-    "FA", # flake8-future-annotations
-    "ISC", # flake8-implicit-str-concat
-    "PIE", # flake8-pie ( misc lints )
-    "PERF", # Perflint
-    "PTH", # flake8-use-pathlib
-    "SIM", # flake8-simplify
-    "SLOT", # flake8-slots
-    "TD", # flake8-todos (EXCLUDE TD002)
-    "TID", # flake8-tidy-imports
-    "UP", # pyupgrade
-    "RET", # flake8-return
-    "RSE", # flake8-raise (RSE)
+    "BLE",   # flake8-blind-except
+    "C4",    # flake8-comprehensions
+    "C90",   # mccabe (complexity)
+    "FA",    # flake8-future-annotations
+    "ISC",   # flake8-implicit-str-concat
+    "PIE",   # flake8-pie ( misc lints )
+    "PERF",  # Perflint
+    "PTH",   # flake8-use-pathlib
+    "SIM",   # flake8-simplify
+    "SLOT",  # flake8-slots
+    "TD",    # flake8-todos (EXCLUDE TD002)
+    "TID",   # flake8-tidy-imports
+    "UP",    # pyupgrade
+    "RET",   # flake8-return
+    "RSE",   # flake8-raise (RSE)
 
     # Typing
     # "ANN", # flake8-annotations
