@@ -1,8 +1,31 @@
 import ast
 from pathlib import Path
+from enum import StrEnum
+from collections.abc import Generator
 
 
-def auto_detect(path: Path):
+class LazyEntity(StrEnum):
+    LazyObject = "lazy-objects"
+    LazyImports = "lazy-imports"
+    LazyExporter = "lazy-exporter"
+
+
+def auto_detect(paths: list[str | Path] | str | Path) -> dict[LazyEntity, set[str]]:
+    if not isinstance(paths, list):
+        paths = [paths]
+
+    entities = (
+        entity for path in paths for entity in auto_detect_from_path(Path(path))
+    )
+
+    results = {}
+    for entity_type, entity in entities:
+        results.setdefault(entity_type, set()).add(entity)
+
+    return results
+
+
+def auto_detect_from_path(path: Path) -> Generator[tuple[LazyEntity, str], None, None]:
     module_paths = (path,) if path.is_file() else path.glob("**/*.py")
 
     for module_path in module_paths:
@@ -17,15 +40,17 @@ def auto_detect(path: Path):
         yield from from_module_content(name, content)
 
 
-def from_module_content(fullname: str, content: str):
+def from_module_content(
+    fullname: str, content: str
+) -> Generator[tuple[LazyEntity, str], None, None]:
     tree = ast.parse(content)
 
     for with_body in with_from_tree(tree):
-        yield f"~{fullname}"
+        yield LazyEntity.LazyExporter, fullname
         yield from imports_from_tree(fullname, with_body)
 
 
-def with_from_tree(tree: ast.AST):
+def with_from_tree(tree: ast.AST) -> Generator[ast.stmt]:
     for node in ast.walk(tree):
         if isinstance(node, ast.With):
             if not any(is_lazy_import(item.context_expr) for item in node.items):
@@ -34,7 +59,7 @@ def with_from_tree(tree: ast.AST):
             yield from node.body
 
 
-def is_lazy_import(node: ast.AST):
+def is_lazy_import(node: ast.AST) -> bool:
     if not isinstance(node, ast.Call):
         return False
 
@@ -42,6 +67,7 @@ def is_lazy_import(node: ast.AST):
         return False
 
     func = node.func
+
     return (isinstance(func, ast.Name) and func.id == "lazy_imports") or (
         isinstance(func, ast.Attribute)
         and isinstance(func.value, ast.Name)
@@ -50,37 +76,22 @@ def is_lazy_import(node: ast.AST):
     )
 
 
-def imports_from_tree(fullname: str, tree: ast.AST):
+def imports_from_tree(
+    fullname: str, tree: ast.AST
+) -> Generator[tuple[LazyEntity, str], None, None]:
     parts = fullname.split(".")
 
     for node in ast.walk(tree):
         if isinstance(node, ast.Import):
-            yield from (n.name for n in node.names)
+            yield from ((LazyEntity.LazyImports, n.name) for n in node.names)
         elif isinstance(node, ast.ImportFrom):
             if (level := node.level) > 1:
-                module = ".".join(parts[: 1 - level] + [node.module])
+                module = ".".join([*parts[: 1 - level], node.module])
             elif level == 1:
                 module = fullname + "." + node.module
             else:
                 module = node.module
 
-            yield from (f"{module}:{n.name}#2" for n in node.names)
-
-
-def main():
-    import argparse
-    from . import LazyModules
-
-    parser = argparse.ArgumentParser()
-    parser.add_argument("path", type=Path)
-    args = parser.parse_args()
-
-    lazy_modules = LazyModules()
-
-    lazy_modules.update(auto_detect(args.path))
-
-    print(",".join(lazy_modules))
-
-
-if __name__ == "__main__":
-    main()
+            yield from (
+                (LazyEntity.LazyObject, f"{module}:{n.name}") for n in node.names
+            )
